requirements.txt
asgiref==3.8.1
Django==5.1.7
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.2
gunicorn
djangorestframework==3.15.0
django-filter==24.1
markdown==3.5.2
